MOVE] src/domain/auth -> src/auth
controller 가 없어졌으므로 더이상 domain이 아님
