+++ src/auth/auth.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
-import { PrismaModule } from '../../prisma/prisma.module';
+import { PrismaModule } from '../prisma/prisma.module';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import { JwtStrategy } from '../../common/strategy/jwt.strategy';
-import { CryptoModule } from '../../crypto/crypto.module';
+import { JwtStrategy } from '../common/strategy/jwt.strategy';
+import { CryptoModule } from '../crypto/crypto.module';
 
 @Module({
 	controllers: [],
