+++ src/domain/user/user.module.ts
@@ -3,7 +3,7 @@ import { UserController } from './user.controller';
 import { UserService } from './user.service';
 import { PrismaModule } from '../../prisma/prisma.module';
 import { CryptoModule } from 'src/crypto/crypto.module';
-import { AuthModule } from '../auth/auth.module';
+import { AuthModule } from '../../auth/auth.module';
 
 @Module({
 	imports: [PrismaModule, CryptoModule, AuthModule],
